Тройной тернарник
