+++ src/js/eventModification.js
@@ -1,6 +1,5 @@
-import { isRetina } from './isRetina.js';
-
 export let arrEvents;
+let isRetina;
 let card;
 const increment = () => card += 1;  
 
@@ -9,8 +8,6 @@ export const eventsModif = (events) => {
   arrEvents = [];
   card = -1; 
   return events.map(event => {
-    console.log(`event-${card+1} :>> `, event);
-
     const eventNew = {
       city: findCity(event),
       country: findCountry(event),
@@ -27,7 +24,6 @@ export const eventsModif = (events) => {
       timeZone: findTimeZone(event),
       url: event.url ? event.url : ''
     }
-    console.log(`eventNew :>> `, eventNew);
     arrEvents.push(eventNew);
     return eventNew;
   }
@@ -53,15 +49,15 @@ const findLocalDate = (event) => event.dates.start.localDate ? event.dates.start
 const findLocalTime = (event) => event.dates.start.localTime ? (event.dates.start.localTime).slice(0, 5) : 
   event._embedded.venues[0].start ? (event._embedded.venues[0].start.localTime).slice(0, 5) : '';
 
-const findPlace = (event) => event._embedded ? event._embedded.venues[0].name :''
+const findPlace = (event) => event._embedded ? event._embedded.venues[0].name : 5>8 ? 'hhhhh' : ''
 
 function findPricesBySt (prices) {
-  if (!prices) {return}
+  if (!prices) {return ''}
   return prices.find(price => price.type === "standard");
 };
 
 function findPricesByVip (prices) {
-  if (!prices) {return}
+  if (!prices) {return ''}
   return prices.find(price => price.type === "vip");
 };
 
@@ -69,9 +65,21 @@ const findTimeZone = (event) => event.dates.timezone ? event.dates.timezone :
   event._embedded.venues[0].timezone ? event._embedded.venues[0].timezone : '';
 
 function filterEventsByPriceRangeVip (events) {
-  if (!events) {return}
+  if (!events) {return ''}
   return events
           .filter(event => event.priceRanges
           .some(priceRang => priceRang.type === "vip"))
       
 };
+
+// document.addEventListener("DOMContentLoaded", isRetinaScreen);
+window.addEventListener("DOMContentLoaded", isRetinaScreen);
+
+function isRetinaScreen(){/*проверка Retina*/
+    isRetina = ((window
+      .matchMedia && (window
+        .matchMedia('only screen and (min-resolution: 192dpi), only screen and (min-resolution: 2dppx), only screen and (min-resolution: 75.6dpcm)')
+        .matches || window
+        .matchMedia('only screen and (-webkit-min-device-pixel-ratio: 2), only screen and (-o-min-device-pixel-ratio: 2/1), only screen and (min--moz-device-pixel-ratio: 2), only screen and (min-device-pixel-ratio: 2)').matches)) || (window.devicePixelRatio && window.devicePixelRatio >= 2)) && /(iPad|iPhone|iPod)/g
+        .test(navigator.userAgent);
+   }
