+++ src/js/modalClose.js
@@ -1,25 +1,16 @@
 import refs from './refs.js';
 
 refs.buttonClose.addEventListener("click", onCloseMainModal);
-refs.overlay.addEventListener("click", onBackdropClick);
+refs.overlay.addEventListener("click", closeMainModalClickOrESC);
+document.addEventListener("keydown", closeMainModalClickOrESC);
 
-document.addEventListener("keydown",
-
-function closeMainModalESC(e) {
-    if (e.key === "Escape") {
-    onCloseMainModal();
-    }
-});
-
-function onCloseMainModal() {
+export default function onCloseMainModal() {
     refs.overlay.classList.add("is-hidden");
     document.body.classList.remove("overlay-show")
 }
 
-function onBackdropClick(e) {
-    if (e.currentTarget === e.target) {
-      onCloseMainModal();
+function closeMainModalClickOrESC(e) {
+    if (e.key === "Escape" || e.currentTarget === e.target) {
+    onCloseMainModal();
     }
 }
-
-export default onCloseMainModal;
